Increase HTTP server backlog size to 2
In Google Chrome on macOS, this fixes random `ERR_SOCKET_NOT_CONNECTED`
errors when using browser-based credential providers.

The error appears to be related to Chrome's "Preload pages" feature,
located here: chrome://settings/preloading.

It causes Chrome to make an extra socket connection that frequently (but
not always) occupies the existing backlog size of `1` and causes the
subsequent `/redshift/` HTTP request to fail with
`ERR_SOCKET_NOT_CONNECTED`.

The error does not happen if preloading is disabled.

Since preloading is enabled by default, I thought it would be better to
increase the backlog size than to ask the user to disable preloading.

Increasing the backlog size to `2` seems to be enough to accommodate the
extra socket connection for "preloading" and allow the `/redshift/`
request to succeed.